+++ session.go
@@ -2,44 +2,44 @@ package transmission
 
 // SetSessionArgs arguments for Session.Set
 type SetSessionArgs struct {
-	AltSpeedDown              int     `json:"alt-speed-down,omitempty"`
+	AltSpeedDown              int64     `json:"alt-speed-down,omitempty"`
 	AltSpeedEnabled           bool    `json:"alt-speed-enabled,omitempty"`
-	AltSpeedTimeBegin         int     `json:"alt-speed-time-begin,omitempty"`
+	AltSpeedTimeBegin         int64     `json:"alt-speed-time-begin,omitempty"`
 	AltSpeedTimeEnabled       bool    `json:"alt-speed-time-enabled,omitempty"`
-	AltSpeedTimeEnd           int     `json:"alt-speed-time-end,omitempty"`
-	AltSpeedTimeDay           int     `json:"alt-speed-time-day,omitempty"`
-	AltSpeedUp                int     `json:"alt-speed-up,omitempty"`
+	AltSpeedTimeEnd           int64     `json:"alt-speed-time-end,omitempty"`
+	AltSpeedTimeDay           int64     `json:"alt-speed-time-day,omitempty"`
+	AltSpeedUp                int64     `json:"alt-speed-up,omitempty"`
 	BlocklistURL              string  `json:"blocklist-url,omitempty"`
 	BlocklistEnabled          bool    `json:"blocklist-enabled,omitempty"`
-	CacheSizeMb               int     `json:"cache-size-mb,omitempty"`
+	CacheSizeMb               int64     `json:"cache-size-mb,omitempty"`
 	DownloadDir               string  `json:"download-dir,omitempty"`
-	DownloadQueueSize         int     `json:"download-queue-size,omitempty"`
+	DownloadQueueSize         int64     `json:"download-queue-size,omitempty"`
 	DownloadQueueEnabled      bool    `json:"download-queue-enabled,omitempty"`
 	DhtEnabled                bool    `json:"dht-enabled,omitempty"`
 	Encryption                string  `json:"encryption,omitempty"`
-	IdleSeedingLimit          int     `json:"idle-seeding-limit,omitempty"`
+	IdleSeedingLimit          int64     `json:"idle-seeding-limit,omitempty"`
 	IdleSeedingLimitEnabled   bool    `json:"idle-seeding-limit-enabled,omitempty"`
 	IncompleteDir             string  `json:"incomplete-dir,omitempty"`
 	IncompleteDirEnabled      bool    `json:"incomplete-dir-enabled,omitempty"`
 	LpdEnabled                bool    `json:"lpd-enabled,omitempty"`
-	PeerLimitGlobal           int     `json:"peer-limit-global,omitempty"`
-	PeerLimitPerTorrent       int     `json:"peer-limit-per-torrent,omitempty"`
+	PeerLimitGlobal           int64     `json:"peer-limit-global,omitempty"`
+	PeerLimitPerTorrent       int64     `json:"peer-limit-per-torrent,omitempty"`
 	PexEnabled                bool    `json:"pex-enabled,omitempty"`
-	PeerPort                  int     `json:"peer-port,omitempty"`
+	PeerPort                  int64     `json:"peer-port,omitempty"`
 	PeerPortRandomOnStart     bool    `json:"peer-port-random-on-start,omitempty"`
 	PortForwardingEnabled     bool    `json:"port-forwarding-enabled,omitempty"`
 	QueueStalledEnabled       bool    `json:"queue-stalled-enabled,omitempty"`
-	QueueStalledMinutes       int     `json:"queue-stalled-minutes,omitempty"`
+	QueueStalledMinutes       int64     `json:"queue-stalled-minutes,omitempty"`
 	RenamePartialFiles        bool    `json:"rename-partial-files,omitempty"`
 	ScriptTorrentDoneFilename string  `json:"script-torrent-done-filename,omitempty"`
 	ScriptTorrentDoneEnabled  bool    `json:"script-torrent-done-enabled,omitempty"`
 	SeedRatioLimit            float64 `json:"seedRatioLimit,omitempty"`
 	SeedRatioLimited          bool    `json:"seedRatioLimited,omitempty"`
-	SeedQueueSize             int     `json:"seed-queue-size,omitempty"`
+	SeedQueueSize             int64     `json:"seed-queue-size,omitempty"`
 	SeedQueueEnabled          bool    `json:"seed-queue-enabled,omitempty"`
-	SpeedLimitDown            int     `json:"speed-limit-down,omitempty"`
+	SpeedLimitDown            int64     `json:"speed-limit-down,omitempty"`
 	SpeedLimitDownEnabled     bool    `json:"speed-limit-down-enabled,omitempty"`
-	SpeedLimitUp              int     `json:"speed-limit-up,omitempty"`
+	SpeedLimitUp              int64     `json:"speed-limit-up,omitempty"`
 	SpeedLimitUpEnabled       bool    `json:"speed-limit-up-enabled,omitempty"`
 	StartAddedTorrents        bool    `json:"start-added-torrents,omitempty"`
 	TrashOriginalTorrentFiles bool    `json:"trash-original-torrent-files,omitempty"`
@@ -51,48 +51,48 @@ type SetSessionArgs struct {
 // session and interact with it
 type Session struct {
 	Client                    *Client `json:"-"`
-	AltSpeedDown              int     `jsonn:"alt-speed-down"`
+	AltSpeedDown              int64     `jsonn:"alt-speed-down"`
 	AltSpeedEnabled           bool    `json:"alt-speed-enabled"`
-	AltSpeedTimeBegin         int     `json:"alt-speed-time-begin"`
+	AltSpeedTimeBegin         int64     `json:"alt-speed-time-begin"`
 	AltSpeedTimeEnabled       bool    `json:"alt-speed-time-enabled"`
-	AltSpeedTimeEnd           int     `json:"alt-speed-time-end"`
-	AltSpeedTimeDay           int     `json:"alt-speed-time-day"`
-	AltSpeedUp                int     `json:"alt-speed-up"`
+	AltSpeedTimeEnd           int64     `json:"alt-speed-time-end"`
+	AltSpeedTimeDay           int64     `json:"alt-speed-time-day"`
+	AltSpeedUp                int64     `json:"alt-speed-up"`
 	BlocklistURL              string  `json:"blocklist-url"`
 	BlocklistEnabled          bool    `json:"blocklist-enabled"`
-	BlocklistSize             int     `json:"blocklist-size"`
-	CacheSizeMb               int     `json:"cache-size-mb"`
+	BlocklistSize             int64     `json:"blocklist-size"`
+	CacheSizeMb               int64     `json:"cache-size-mb"`
 	ConfigDir                 string  `json:"config-dir"`
 	DownloadDir               string  `json:"download-dir"`
-	DownloadQueueSize         int     `json:"download-queue-size"`
+	DownloadQueueSize         int64     `json:"download-queue-size"`
 	DownloadQueueEnabled      bool    `json:"download-queue-enabled"`
 	DhtEnabled                bool    `json:"dht-enabled"`
 	Encryption                string  `json:"encryption"`
-	IdleSeedingLimit          int     `json:"idle-seeding-limit"`
+	IdleSeedingLimit          int64     `json:"idle-seeding-limit"`
 	IdleSeedingLimitEnabled   bool    `json:"idle-seeding-limit-enabled"`
 	IncompleteDir             string  `json:"incomplete-dir"`
 	IncompleteDirEnabled      bool    `json:"incomplete-dir-enabled"`
 	LpdEnabled                bool    `json:"lpd-enabled"`
-	PeerLimitGlobal           int     `json:"peer-limit-global"`
-	PeerLimitPerTorrent       int     `json:"peer-limit-per-torrent"`
+	PeerLimitGlobal           int64     `json:"peer-limit-global"`
+	PeerLimitPerTorrent       int64     `json:"peer-limit-per-torrent"`
 	PexEnabled                bool    `json:"pex-enabled"`
-	PeerPort                  int     `json:"peer-port"`
+	PeerPort                  int64     `json:"peer-port"`
 	PeerPortRandomOnStart     bool    `json:"peer-port-random-on-start"`
 	PortForwardingEnabled     bool    `json:"port-forwarding-enabled"`
 	QueueStalledEnabled       bool    `json:"queue-stalled-enabled"`
-	QueueStalledMinutes       int     `json:"queue-stalled-minutes"`
+	QueueStalledMinutes       int64     `json:"queue-stalled-minutes"`
 	RenamePartialFiles        bool    `json:"rename-partial-files"`
-	RPCVersion                int     `json:"rpc-version"`
-	RPCVersionMinimum         int     `json:"rpc-version-minimum"`
+	RPCVersion                int64     `json:"rpc-version"`
+	RPCVersionMinimum         int64     `json:"rpc-version-minimum"`
 	ScriptTorrentDoneFilename string  `json:"script-torrent-done-filename"`
 	ScriptTorrentDoneEnabled  bool    `json:"script-torrent-done-enabled"`
 	SeedRatioLimit            float64 `json:"seedRatioLimit"`
 	SeedRatioLimited          bool    `json:"seedRatioLimited"`
-	SeedQueueSize             int     `json:"seed-queue-size"`
+	SeedQueueSize             int64     `json:"seed-queue-size"`
 	SeedQueueEnabled          bool    `json:"seed-queue-enabled"`
-	SpeedLimitDown            int     `json:"speed-limit-down"`
+	SpeedLimitDown            int64     `json:"speed-limit-down"`
 	SpeedLimitDownEnabled     bool    `json:"speed-limit-down-enabled"`
-	SpeedLimitUp              int     `json:"speed-limit-up"`
+	SpeedLimitUp              int64     `json:"speed-limit-up"`
 	SpeedLimitUpEnabled       bool    `json:"speed-limit-up-enabled"`
 	StartAddedTorrents        bool    `json:"start-added-torrents"`
 	TrashOriginalTorrentFiles bool    `json:"trash-original-torrent-files"`
@@ -103,32 +103,32 @@ type Session struct {
 
 // Statictics represent session statictics
 type Statictics struct {
-	ActiveTorrentCount int
-	DownloadSpeed      int
-	PausedTorrentCount int
-	TorrentCount       int
-	UploadSpeed        int
+	ActiveTorrentCount int64
+	DownloadSpeed      int64
+	PausedTorrentCount int64
+	TorrentCount       int64
+	UploadSpeed        int64
 	CumulativeStats    *StaticticDetail `json:"cumulative-stats"`
 	CurrentStats       *StaticticDetail `json:"current-stats"`
 }
 
 // StaticticDetail represent statictics details
 type StaticticDetail struct {
-	UploadedBytes   int
-	DownloadedBytes int
-	FilesAdded      int
-	SessionCount    int
-	SecondsActive   int
+	UploadedBytes   int64
+	DownloadedBytes int64
+	FilesAdded      int64
+	SessionCount    int64
+	SecondsActive   int64
 }
 
 // Units in session
 type Units struct {
 	SpeedUnits  []string `json:"speed-units"`
-	SpeedBytes  int      `json:"speed-bytes"`
+	SpeedBytes  int64      `json:"speed-bytes"`
 	SizeUnits   []string `json:"size-units"`
-	SizeBytes   int      `json:"size-bytes"`
+	SizeBytes   int64      `json:"size-bytes"`
 	MemoryUnits []string `json:"memory-units"`
-	MemoryBytes int      `json:"memory-bytes"`
+	MemoryBytes int64      `json:"memory-bytes"`
 }
 
 // Set set session params see SetSessionArgs
